+++ companion/manifest.json
@@ -18,7 +18,7 @@
 		"type": "node18",
 		"api": "nodejs-ipc",
 		"apiVersion": "0.0.0",
-		"entrypoint": "../osc.js"
+		"entrypoint": "../src/main.js"
 	},
 	"manufacturer": "Songbeamer",
 	"products": ["OSC"],
