+++ package.json
@@ -1,7 +1,8 @@
 {
 	"name": "songbeamer-osc",
 	"version": "2.0.0",
-	"main": "osc.js",
+	"main": "src/main.js",
+	"type": "module",
 	"scripts": {
 		"format": "prettier --write ."
 	},
@@ -20,5 +21,6 @@
 		"osc": "^2.4.1",
 		"utf-8-validate": "^6.0.3",
 		"ws": "^8.13.0"
-	}
+	},
+	"prettier": "@companion-module/tools/.prettierrc.json"
 }
