+++ src/variables.js
@@ -1,0 +1,26 @@
+export const variables = [
+	{
+		name: 'presentation state',
+		variableId: 'presentation_state',
+	},
+	{
+		name: 'presentation page',
+		variableId: 'presentation_page',
+	},
+	{
+		name: 'playlist itemindex (starting with 1)',
+		variableId: 'playlist_itemindex',
+	},
+	{
+		name: 'Presentation message text',
+		variableId: 'presentation_message_text',
+	},
+	{
+		name: 'Presentation message visibility',
+		variableId: 'presentation_message_visible',
+	},
+	{
+		name: 'presentation pagecount',
+		variableId: 'presentation_pagecount',
+	},
+]
