--- upgrades.js
@@ -1,10 +1,0 @@
-module.exports = [
-	function (context, props) {
-		// This is a placeholder than now cannot be used/removed
-		return {
-			updatedConfig: null,
-			updatedActions: [],
-			updatedFeedbacks: [],
-		}
-	},
-]
